+++ requery-android/src/main/java/io/requery/android/EntityParceler.java
@@ -60,7 +60,7 @@ public class EntityParceler<T> {
             }
             PropertyState state = PropertyState.LOADED;
             if (!type.isStateless()) {
-                state = PropertyState.valueOf(in.readValue(null).toString());
+                state = PropertyState.valueOf(in.readString());
             }
             proxy.setObject(attribute, value, state);
         }